# docs/api/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oulib API Documentation",
    "description": "API documentation for Oulib - the library management system",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lấy danh sách người dùng",
        "description": "Trả về danh sách người dùng có phân trang cho SYSADMIN hoặc LIBRARIAN",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Số trang, bắt đầu từ 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số phần tử mỗi trang",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách người dùng thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseUserResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Đăng ký tài khoản người dùng",
        "description": "Tạo mới tài khoản người dùng với vai trò USER từ thông tin đăng ký cơ bản",
        "operationId": "createUser",
        "requestBody": {
          "description": "Thông tin đăng ký tài khoản người dùng",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tạo người dùng thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu đăng ký không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "409": {
            "description": "Người dùng đã tồn tại",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/staff": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Tạo tài khoản nhân sự",
        "description": "Cho phép SYSADMIN tạo tài khoản LIBRARIAN hoặc SYSADMIN mới",
        "operationId": "createStaff",
        "requestBody": {
          "description": "Thông tin tài khoản nhân sự cần tạo",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tạo tài khoản nhân sự thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "409": {
            "description": "Email đã được sử dụng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{id}/status": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Cập nhật trạng thái người dùng",
        "description": "Cập nhật trạng thái tài khoản theo userId cho SYSADMIN hoặc LIBRARIAN",
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID của người dùng cần cập nhật trạng thái",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Trạng thái mới của tài khoản người dùng",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cập nhật trạng thái thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu trạng thái không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền thực hiện",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lấy hồ sơ người dùng hiện tại",
        "description": "Lấy thông tin hồ sơ của tài khoản đang đăng nhập dựa trên JWT",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Lấy hồ sơ thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Cập nhật hồ sơ cá nhân",
        "description": "Cập nhật thông tin hồ sơ của tài khoản hiện tại bằng dữ liệu multipart",
        "operationId": "updateMyProfile",
        "requestBody": {
          "description": "Dữ liệu multipart gồm phần data chứa thông tin hồ sơ và phần avatar tùy chọn",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "$ref": "#/components/schemas/UserUpdateRequest"
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Tệp ảnh đại diện tải lên, có thể bỏ trống"
                  }
                },
                "required": [
                  "data"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cập nhật hồ sơ thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu cập nhật không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Đổi mật khẩu tài khoản hiện tại",
        "description": "Người dùng đã đăng nhập đổi mật khẩu bằng cách cung cấp mật khẩu cũ và mật khẩu mới",
        "operationId": "changePassword",
        "requestBody": {
          "description": "Thông tin mật khẩu cũ và mật khẩu mới",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Đổi mật khẩu thành công",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu mật khẩu không hợp lệ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực hoặc mật khẩu cũ không đúng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi hệ thống",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "description": "DTO phản hồi UserResponse",
        "properties": {
          "id": {
            "type": "string",
            "description": "Trường id",
            "example": "id-001"
          },
          "username": {
            "type": "string",
            "description": "Trường username",
            "example": "Nguyen Van A"
          },
          "fullName": {
            "type": "string",
            "description": "Trường fullName",
            "example": "Nguyen Van A"
          },
          "email": {
            "type": "string",
            "description": "Trường email",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "description": "Trường role",
            "example": "LIBRARIAN"
          },
          "avatar": {
            "type": "string",
            "description": "Trường avatar",
            "example": "https://example.com/resource"
          },
          "status": {
            "type": "string",
            "description": "Trường status",
            "example": "ACTIVE"
          },
          "active": {
            "type": "boolean",
            "description": "Trường active",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "description": "Trường createdAt",
            "example": "sample"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "UserCreationRequest": {
        "type": "object",
        "description": "DTO yêu cầu UserCreationRequest",
        "properties": {
          "email": {
            "type": "string",
            "description": "Trường email",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Trường fullName",
            "example": "Nguyen Van A",
            "maxLength": 50,
            "minLength": 1
          },
          "password": {
            "type": "string",
            "description": "Trường password",
            "example": "Password@123",
            "maxLength": 2147483647,
            "minLength": 6
          }
        },
        "required": [
          "email",
          "fullName",
          "password"
        ]
      },
      "StaffCreationRequest": {
        "type": "object",
        "description": "DTO yêu cầu StaffCreationRequest",
        "properties": {
          "email": {
            "type": "string",
            "description": "Trường email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Trường password",
            "example": "Password@123",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "fullName": {
            "type": "string",
            "description": "Trường fullName",
            "example": "Nguyen Van A",
            "maxLength": 50,
            "minLength": 1
          },
          "role": {
            "type": "string",
            "description": "Trường role",
            "enum": [
              "SYSADMIN",
              "LIBRARIAN",
              "USER"
            ],
            "example": "LIBRARIAN"
          }
        },
        "required": [
          "email",
          "fullName",
          "password",
          "role"
        ]
      },
...
      "UserStatusUpdateRequest": {
        "type": "object",
        "description": "DTO yêu cầu UserStatusUpdateRequest",
        "properties": {
          "status": {
            "type": "string",
            "description": "Trường status",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "DELETED",
              "PRIVATE"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "status"
        ]
      },
      "ApiResponseUserResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
      "UserUpdateRequest": {
        "type": "object",
        "description": "Dữ liệu hồ sơ người dùng cần cập nhật",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Trường fullName",
            "example": "Nguyen Van A",
            "maxLength": 50,
            "minLength": 1
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "description": "DTO yêu cầu ChangePasswordRequest",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "Trường oldPassword",
            "example": "Password@123"
          },
          "newPassword": {
            "type": "string",
            "description": "Trường newPassword",
            "example": "Password@123",
            "maxLength": 2147483647,
            "minLength": 6
          }
        },
        "required": [
          "newPassword",
          "oldPassword"
        ]
      },
...
      "ApiResponsePageResponseUserResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseUserResponse"
          }
        }
      },
      "PageResponseUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}